extprocplugin: Add iqbin file support.
Convert data types from double to float throughout the pipeline for better
performance.
Add IQBinHeader struct and support for reading/writing IQ binary files
with headers.

+++ packages/extproc/plugins/extprocplugin/include/extprocplugin/extprocinstrument.h
@@ -51,7 +51,7 @@ Q_SIGNALS:
 	void requestAnalysisInfo(const QString &type);
 	void outputConfigured(const OutputConfig &outConfig);
 	void analysisConfigChanged(const QString &type, const QVariantMap &config);
-	void bufferDataReady(QVector<QVector<double>> &inputData);
+	void bufferDataReady(QVector<QVector<float>> &inputData);
 	void runPressed(bool en);
 	void requestNewData();
